Add files via upload
Finance Calculator - Python
This project was built with the goal of providing the user with a ﬁnancial calculator. This calculator was built using Python and has 2 choices the user can make. The user may choose between calculating an investment's value of x amount of years or calculating their monthly bond repayment.